+++ lib/occams-record/eager_loaders/base.rb
@@ -15,11 +15,13 @@ module OccamsRecord
       # @param ref [ActiveRecord::Association] the ActiveRecord association
       # @param scope [Proc] a scope to apply to the query (optional)
       # @param use [Array(Module)] optional Module to include in the result class (single or array)
+      # @param as [Symbol] Load the association into this attr instead
       # @param eval_block [Proc] a block where you may perform eager loading on *this* association (optional)
       #
-      def initialize(ref, scope = nil, use = nil, &eval_block)
-        @ref, @scope, @use, @eval_block = ref, scope, use, eval_block
-        @name, @model = ref.name.to_s, ref.klass
+      def initialize(ref, scope = nil, use: nil, as: nil, &eval_block)
+        @ref, @scope, @use, @as, @eval_block = ref, scope, use, as, eval_block
+        @model = ref.klass
+        @name = (as || ref.name).to_s
       end
 
       #