+++ lib/occams-record/query.rb
@@ -75,6 +75,16 @@ module OccamsRecord
     alias_method :to_a, :run
 
     #
+    # Run the query and return the first result (which could be nil). This WILL append a LIMIT 1 to the query.
+    #
+    # @return [OccamsRecord::Results::Row]
+    #
+    def first
+      scope.limit! 1
+      run[0]
+    end
+
+    #
     # If you pass a block, each result row will be yielded to it. If you don't,
     # an Enumerable will be returned.
     #